+++ ImmoLiteFramework/Assets/ImmoLiteFramework/CHANGELOG.md
@@ -3,4 +3,10 @@
 ## [0.0.1] - 2026-01-22
 
 ### Added
-- Basic module for a simple but solid game framework: event, entity, finite state machine, and resource modules.
+- Basic managers: event, UI, and resource
+
+
+## [0.0.2] - 2026-01-22
+
+### Added
+- Support `GameObject` creation menu item to quickly create a ready framework object